fix(sqllab): reorganize extension tests per reviewer feedback
Move extensionTestHelpers to spec/helpers (not SQL Lab specific),
relocate error boundary test to StatusBar.test.tsx, move PanelToolbar
click test to its own component test file, and remove duplicate
dispose test already covered by core/views/index.test.ts.
