+++ superset-frontend/src/SqlLab/components/QueryHistory/QueryHistory.test.tsx
@@ -29,7 +29,7 @@ import { ViewLocations } from 'src/SqlLab/contributions';
 import {
   registerToolbarAction,
   cleanupExtensions,
-} from 'src/SqlLab/test-utils/extensionTestHelpers';
+} from 'spec/helpers/extensionTestHelpers';
 
 const mockedProps = {
   queryEditorId: defaultQueryEditor.id,
