+++ superset-frontend/src/SqlLab/components/TableExploreTree/TableExploreTree.test.tsx
@@ -26,7 +26,7 @@ import { ViewLocations } from 'src/SqlLab/contributions';
 import {
   registerToolbarAction,
   cleanupExtensions,
-} from 'src/SqlLab/test-utils/extensionTestHelpers';
+} from 'spec/helpers/extensionTestHelpers';
 import TableExploreTree from '.';
 
 jest.mock(
